playground_groq/worker/requirements.txt
groq
openai-whisper
resemblyzer
webrtcvad
pydub
requests
beautifulsoup4
torch
transformers
